sepparate types from code
