+++ .travis.yml
@@ -7,6 +7,13 @@ sudo: true
 
 env:
   - CXX=g++-4.8 NODE_ENV=testing
+
+services:
+  - postgresql
+
+before_script:
+  - psql -c 'create database travis_ci_test;' -U postgres
+
 addons:
   apt:
     sources:
